+++ src/main/java/com/marsspiders/ukwa/solr/data/FacetFields.java
@@ -12,6 +12,9 @@ public class FacetFields {
     @JsonProperty("public_suffix")
     private LinkedList<String> publicSuffixes;
 
+    @JsonProperty("type")
+    private LinkedList<String> types;
+
     @JsonProperty("content_type_norm")
     private LinkedList<String> contentTypes;
 
@@ -24,6 +27,9 @@ public class FacetFields {
     @JsonProperty("collection")
     private LinkedList<String> collections;
 
+    @JsonProperty("access_terms")
+    private LinkedList<String> accessTerms;
+
     public LinkedList<String> getPublicSuffixes() {
         return publicSuffixes;
     }
@@ -32,12 +38,20 @@ public class FacetFields {
         this.publicSuffixes = publicSuffixes;
     }
 
+    public LinkedList<String> getTypes() {
+        return types;
+    }
+
     public LinkedList<String> getContentTypes() {
         return contentTypes;
     }
 
     public void setContentTypes(LinkedList<String> contentTypes) {
         this.contentTypes = contentTypes;
+    }
+
+    public void setTypes(LinkedList<String> types) {
+        this.types = types;
     }
 
     public LinkedList<String> getHosts() {
@@ -62,6 +76,14 @@ public class FacetFields {
 
     public void setCollections(LinkedList<String> collections) {
         this.collections = collections;
+    }
+
+    public LinkedList<String> getAccessTerms() {
+        return accessTerms;
+    }
+
+    public void setAccessTerms(LinkedList<String> accessTerms) {
+        this.accessTerms = accessTerms;
     }
 
     @Override