+++ src/main/resources/static/ukwa/includes/common.js
@@ -1,49 +1,37 @@
 //common on load functions
 
-function expandCollDescript(ev, object) {
-
-	ev.preventDefault(); //don't activate link to collection
-	
-	$(".img-square-caption").addClass("shadow"); //add shadow to all elements in case some is missing
-	parentElem=object.parent().parent();
-	$(".collection-description").height(parentElem.height());
-	$(".collection-description > .collection-description-content").text(object.parent().attr("data-descript"));
-	$(".collection-description").css("width", parentElem.width()+40);
-	$(".collection-description").css("top", parentElem.offset().top);
-	$(".collection-description").css("left", parentElem.offset().left);
-	$(".collection-description").css("display", "block");
-	parentElem.removeClass("shadow");
-	if ($(".collection-description-content").height()>parentElem.height()) {
-		$(".collection-description").height($(".collection-description-content").height());
-	} else {
-		$(".collection-description").height(parentElem.height());
-	}
-	$(".collection-readmore-up").focus();
-	
-}
-
-function collapseCollDescript(ev, object) {
-
-	ev.preventDefault(); //don't activate link to collection
-	$(".collection-description").css("display", "none");
-	$(".img-square-caption").addClass("shadow");	
-	object.focus();
-	
-}
-
 function checkboxSize() {
 	$(".form-check-cont").each(function(index, element) {
-        var lbl=$(this).find("label");
-		$(this).height(lbl.height()+10);
+		$(this).height($(this).find("label").height()+10);
     });	
+}
+
+function rearangeQuestions() {
+	
+	$(".q-description").each(function(index, element) {
+		//if mobile
+		if ($(window).width()<768) {
+			var elem=$(this).detach();
+			elem.insertAfter($(".q-question[data-descriptid='"+$(this).attr("id")+"']"));		
+		} else {	
+			var id=parseInt($(this).attr("id"));
+			var target=Math.ceil(id / 3)*3;
+			var elem=$(this).detach();
+			elem.insertAfter($(".q-question[data-descriptid='"+target+"']").parent());	
+		}
+    });
+	
 }
 
 $(document).ready(function(e) {
 	
 	//bootstrap tooltips
 	$('[data-toggle="tooltip"]').tooltip(); 
-	$('[data-toggle="tooltip"]').click(function(e) { e.stopPropagation(); });
+	$('[data-toggle="tooltip"]').click(function(e) { e.stopPropagation(); }) //stopps click on the "?" button from expanding/collapsing the filters;
 	
+	//remove tabindex from radio/check boxes
+	$(this).find("input[type=radio], input[type=checkbox]").attr("tabindex", "-1");
+		
 	//radio and check button keyboard
     $(".form-check-cont").each(function(index, element) {
         $(this).on("keypress", function(e) {
@@ -54,52 +42,30 @@ $(document).ready(function(e) {
 	
 	//main menu
 	$(".main-menu-button").click(function(e) {
-        $(".main-menu").addClass("active");
-		$(".main-menu-block").addClass("active");
+        $(".main-menu, .main-menu-block").addClass("active");
 		$("html").css("overflow", "hidden");
     });
 	
 	$(".main-menu-close").click(function(e) {
-        $(".main-menu").removeClass("active");
-		$(".main-menu-block").removeClass("active");
+        $(".main-menu, .main-menu-block").removeClass("active");
 		$("html").css("overflow", "auto");
     });
 	
 	//sidebar collapse
 	$("#toggle-sidebar").click(function(e) {
 		if ($(this).hasClass("open")) {
-			$(this).addClass("closed");
-			$(this).removeClass("open");
+			$(this).addClass("closed").removeClass("open");
 			$(".sidebar-collapse").addClass("open");
 		} else {
-			$(this).addClass("open");
-			$(this).removeClass("closed");
+			$(this).addClass("open").removeClass("closed");
 			$(".sidebar-collapse").removeClass("open");
 		}
 	});
 	
 	//scroll top
-	$(".up-button").click(function(e) {
-		$("html, body").animate({ scrollTop: 0 }, 600);
+	$(".up-button").on('click keyup', function(e) {
+		if(e.which == 13 || e.which == 1) $("html, body").animate({ scrollTop: 0 }, 600);
 	});
-	
-	$(".up-button").keyup(function(e) {
-		if(e.which == 13) $("html, body").animate({ scrollTop: 0 }, 600);
-	});
-	
-	//about video
-	$("#play-about-video").click(function(e) {
-		$("html, body").animate({ scrollTop: "0px" });
-        $(".about-full-video-container").fadeIn("slow");
-		$("#about-full-video")[0].play();
-		$("html").css("overflow","hidden");
-    });
-	
-	$("#close-about-video").click(function(e) {
-        $(".about-full-video-container").fadeOut("slow");
-		$("#about-full-video")[0].pause();
-		$("html").css("overflow","auto");
-    });
 	
 	//replace broken images for collections
 	$(".coll-img").each(function(index, element) {
@@ -111,19 +77,14 @@ $(document).ready(function(e) {
 			console.log(old_img+' replaced with default image.');
 		})
     });
-
-	
-	//match height for collection headings and descriptions
-	$('.img-square-caption').matchHeight();
-	$('.collection-heading').matchHeight();	
 	
 	
 	//add default HTML document type filter
 	$("#search_form").submit(function(e) {
-		$(this).append('<input type="hidden" name="content_type" value="html" />');
+		$(this).append('<input type="hidden" name="content_type" value="Web Page" />');
     });
 	
-	//FAQ and technical
+	//FAQ 
 	$(".q-question").each(function(index, element) {
        	$(this).click(function(e) {
 			e.preventDefault();
@@ -132,27 +93,67 @@ $(document).ready(function(e) {
 			$(this).addClass("active");
 			$("#"+$(this).attr("data-descriptid")).show();
 			$('html, body').animate({
-				scrollTop: $("#"+$(this).attr("data-descriptid")).offset().top
+				scrollTop: $(this).offset().top
 			}, 500);
 		}); 
     });
 	
-	//cookies
+	//cookies notice
 	if (typeof $.cookie('cookies_accepted') === 'undefined' || $.cookie('cookies_accepted')!=="true") {
 		$(".cookies-cont").show();
 	} 
 	
+	//close cookie notice
 	$("#btn_cookies").click(function(e) {
         $.cookie("cookies_accepted", "true", { expires: 365, path: '/' });
 		$(".cookies-cont").hide();
     });
 	
-	//checkboxes and radios
-	checkboxSize();
-	$(window).resize(function(e) {
-        checkboxSize();
+	//on load and resize
+	$(window).on ("load resize", function(e) {
+		checkboxSize(); //expand checkbox size to fit label content
+		rearangeQuestions(); //rearange questions for mobile and desktop
     });
 	
+	//search highlighting
+	if ($(".results-container").length!==0) {
 	
+		//every text section that needs highlighting has this class .results-for-hightlight
+		$(".results-for-highlight").each(function(index, element) {
+		
+			var restext=$(this).html();
+			var terms=$(".main-search-field").val().split(" ");
+			var operators=["AND", "OR", "NOT", "and", "or", "not", "And", "Or", "Not"];
+			var matches=[];
+			var uniqueMatches=[];
+			
+			//match all search terms
+			$.each(terms, function(i, term) {
+				var rexp = new RegExp(term, 'gi');
+				var current_match = restext.match(rexp);
+				matches=matches.concat(current_match)
+			});
+			
+			//remove duplicates
+			$.each(matches, function(i, el){
+				if($.inArray(el, uniqueMatches) === -1) uniqueMatches.push(el);
+			});
+			
+			
+			//replace with highlight style
+			$.each(uniqueMatches, function(i, el){
+				var rexp = new RegExp(el, 'g');
+				if($.inArray(el, operators) === -1) restext=restext.replace(rexp, '<span class="results-highlight">'+el+'</span>');
+			});
+			
+			$(this).html(restext);
+		
+		});
+		
+	}
+	
+	//match height for collection headings and descriptions
+	$('.collection-heading').matchHeight();	
+	$('.img-square-caption').matchHeight();
 	
 });